Data Import, Split, Error%
Changed importData so it just imports all data to data var and n cols

Added splitData that splits the data var into train and test data based on test_size global var. It randomly selects columns for the test and training data that do not overlap.

Added a loop to the main function that runs the SVM multiple times and outputs an average errors per trial and the total % error.